// frontend/src/shared/api/generated/models/createNewsItemRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CodigoActivo.API
 * OpenAPI spec version: 1.0
 */

export interface CreateNewsItemRequest {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  description?: string | null;
  thumbnailId?: string;
}